moved rightclick, added inventory type
Moved OnRightClickClass to a new script, added InventoryType enum to the interfaces for ease of differentiation.
