+++ Assets/Scripts/InventoryUI.cs
@@ -19,6 +19,7 @@ public class InventoryUI : UserInterfaceBase
     public GameObject[] InventorySlots { get { return _inventorySlots; } }
     protected override void PopulateInventory()
     {
+        Type = InterfaceType.Inventory;
         _inventorySlots = new GameObject[Inventory.Inventory.InventoryObject.Length];
 
         for (int i = 0; i < Inventory.Inventory.InventoryObject.Length; i++)
@@ -52,72 +53,3 @@ public class InventoryUI : UserInterfaceBase
 
 }
 
-public class MyRightClickClass : MonoBehaviour, IPointerClickHandler
-{
-    private UIController _mainParent;
-    private InventoryUI _referenceInventory;
-    private EquipScreenUI _referenceEquip;
-
-    private void Start()
-    {
-        _mainParent = transform.parent.parent.GetComponent<UIController>();
-        _referenceInventory = _mainParent.InventoryScreen;
-        _referenceEquip = _mainParent.EquipmentScreen;
-    }
-
-    public void OnPointerClick(PointerEventData eventData)
-    {
-        Debug.Log("click");
-        if (gameObject.GetComponentInParent<InventoryUI>())
-        {
-            Debug.Log("abb");
-            if (eventData.button == PointerEventData.InputButton.Right)
-            {
-                Debug.Log("acc");
-
-                for (int i = 0; i < _referenceEquip.Inventory.Inventory.InventoryObject.Length; i++)
-                {
-                    var obj = _referenceEquip.EquipSlots[i];
-
-                    for (int j = 0; j < _referenceEquip.InventoryUI[obj].AllowedItems.Length; j++)
-                    {
-                        if (_referenceInventory.InventoryUI[gameObject].Item.Id != -1)
-                        {
-                            if (_referenceEquip.InventoryUI[obj].AllowedItems[j] == _referenceInventory.InventoryUI[gameObject].Item.Reference.Type)
-                            {
-                                Debug.Log("i am here");
-                                _referenceInventory.Inventory.MoveItem(_referenceInventory.InventoryUI[gameObject], _referenceEquip.InventoryUI[obj]);
-                                break;
-                            }
-                        }
-                        
-                    }
-                }
-            }
-        }
-        else if (gameObject.GetComponentInParent<EquipScreenUI>())
-        {
-            Debug.Log("amm");
-            if (eventData.button == PointerEventData.InputButton.Right)
-            {
-                Debug.Log("amm2");
-                for (int i = 0; i < _referenceInventory.Inventory.Inventory.InventoryObject.Length; i++)
-                {
-                    Debug.Log("mma3");
-                    var obj = _referenceInventory.InventorySlots[i];
-
-                    if (_referenceInventory.InventoryUI[obj].Item.Id == -1)
-                    {
-                        Debug.Log("i am here too");
-                        _referenceInventory.Inventory.MoveItem(_referenceInventory.InventoryUI[obj], _referenceEquip.InventoryUI[gameObject]);
-                        break;
-                    }
-
-                   
-                }
-            }
-        }
-        
-    }
-
-}
