+++ Assets/Scripts/MyRightClickClass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7980ff516f436444bb1c38ec3c35a557
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
